ActivitySensor fixes: Remove unneeded try/catch error handling, null is returned if there is no sensor use m_ prefix to better distinguish member variables in some cases use C++ standard range based for syntax, not the MS specific "for each" Remove AnyCPU configuration from JS sample that causes build errors
I verified C++ and CS changes with the simulator that Lisa provided, since JS binds to system metadata I was not able to run the code due to the interface change. but this change was simple enough so I'm satisfied with running it up to the point of the exception.